Patch 14
Completamento del frammento details auction e del others profile, correzione di numerosi bug e problemi logici e grafici, nonché completamento della versione 1.0 del Frontend dell'applicazione

+++ LootMarket/app/src/main/java/com/danilo/lootmarket/AuctionViewHistory.kt
@@ -5,4 +5,4 @@ import java.time.LocalDate
 import java.time.ZonedDateTime
 import java.util.Date
 
-data class AuctionViewHistory(val id: Int, val titoloAsta: String, val ultimaOfferta: Float, val dataScadenza: ZonedDateTime, val immagineProdotto: Drawable, val autore: String, val offertaFatta: Boolean)
+data class AuctionViewHistory(val id: Int, val titoloAsta: String, val ultimaOfferta: Double, val dataScadenza: ZonedDateTime, val immagineProdotto: Drawable, val autore: String, val offertaFatta: Boolean)
